"Implemented the Query Annotations and There was an underlying error of the methods not working so also fixed them. "

+++ Day 8/blog_db/src/main/java/com/red/blog_db/service/CommentService.java
@@ -41,15 +41,18 @@ public class CommentService {
     public void createComment(Long myId, Long postId, @NotNull @NonNull CommentEntity commentEntity) throws RuntimeException {
 
 
-            UserEntity user = userRepository.findByUserId(postId);
-            PostEntity post = postRepository.findByPostId(myId);
-            commentEntity.setUserEntity(user);
-            commentEntity.setPostEntity(post);
-            commentRepository.save(commentEntity);
+        UserEntity user = userRepository.findByUserId(postId);
+        PostEntity post = postRepository.findByPostId(myId);
+        commentEntity.setUserEntity(user);
+        commentEntity.setPostEntity(post);
+        commentRepository.save(commentEntity);
 
     }
 
     public void deleteComment(Long myId) {
+        if (!commentRepository.existsById(myId)) {
+            throw new RuntimeException("Comment not found");
+        }
         commentRepository.deleteById(myId);
     }
